tr2/objects: split bandit module
