+++ src/tr2/game/objects/creatures/bandit_1.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void Bandit1_Setup(void);
