+++ src/tr2/game/objects/creatures/bandit_2.c
@@ -1,40 +1,14 @@
-#include "game/objects/creatures/bandit.h"
+#include "game/objects/creatures/bandit_2.h"
 
 #include "game/creature.h"
 #include "game/objects/common.h"
+#include "game/objects/creatures/bandit_common.h"
 #include "global/funcs.h"
 #include "global/vars.h"
 
 #include <libtrx/debug.h>
 
-#define BANDIT1_HITPOINTS 45
-#define BANDIT2_HITPOINTS 50
-#define BANDIT_RADIUS (WALL_L / 10) // = 102
-
-void Bandit1_Setup(void)
-{
-    OBJECT *const obj = Object_GetObject(O_BANDIT_1);
-    if (!obj->loaded) {
-        return;
-    }
-
-    obj->control = Bandit1_Control;
-    obj->collision = Creature_Collision;
-
-    obj->hit_points = BANDIT1_HITPOINTS;
-    obj->radius = BANDIT_RADIUS;
-    obj->shadow_size = UNIT_SHADOW / 2;
-    obj->pivot_length = 0;
-
-    obj->intelligent = 1;
-    obj->save_position = 1;
-    obj->save_hitpoints = 1;
-    obj->save_flags = 1;
-    obj->save_anim = 1;
-
-    g_AnimBones[obj->bone_idx + 6 * 4] |= BF_ROT_Y;
-    g_AnimBones[obj->bone_idx + 8 * 4] |= BF_ROT_Y;
-}
+#define BANDIT_2_HITPOINTS 50
 
 void Bandit2_Setup(void)
 {
@@ -46,7 +20,7 @@ void Bandit2_Setup(void)
     obj->control = Bandit2_Control;
     obj->collision = Creature_Collision;
 
-    obj->hit_points = BANDIT2_HITPOINTS;
+    obj->hit_points = BANDIT_2_HITPOINTS;
     obj->radius = BANDIT_RADIUS;
     obj->shadow_size = UNIT_SHADOW / 2;
     obj->pivot_length = 0;
@@ -76,7 +50,7 @@ void Bandit2B_Setup(void)
     obj->control = Bandit2_Control;
     obj->collision = Creature_Collision;
 
-    obj->hit_points = BANDIT2_HITPOINTS;
+    obj->hit_points = BANDIT_2_HITPOINTS;
     obj->radius = BANDIT_RADIUS;
     obj->shadow_size = UNIT_SHADOW / 2;
     obj->pivot_length = 0;
