+++ src/tr2/game/objects/creatures/bandit_2.h
@@ -1,5 +1,4 @@
 #pragma once
 
-void Bandit1_Setup(void);
 void Bandit2_Setup(void);
 void Bandit2B_Setup(void);
