+++ src/tr2/game/objects/setup.c
@@ -3,7 +3,8 @@
 #include "game/lara/control.h"
 #include "game/lara/hair.h"
 #include "game/objects/common.h"
-#include "game/objects/creatures/bandit.h"
+#include "game/objects/creatures/bandit_1.h"
+#include "game/objects/creatures/bandit_2.h"
 #include "game/objects/creatures/barracuda.h"
 #include "game/objects/creatures/bartoli.h"
 #include "game/objects/creatures/big_eel.h"
