+++ src/tr2/meson.build
@@ -133,7 +133,8 @@ dll_sources = [
   'game/music/music_backend_files.c',
   'game/music/music_main.c',
   'game/objects/common.c',
-  'game/objects/creatures/bandit.c',
+  'game/objects/creatures/bandit_1.c',
+  'game/objects/creatures/bandit_2.c',
   'game/objects/creatures/barracuda.c',
   'game/objects/creatures/bartoli.c',
   'game/objects/creatures/big_eel.c',
